The lists show the read view in the list, so break the UI out.
Each list shows the full item in the list, as if you’d clicked on the
item to view only.  The view only handled in the index, so split out
the ui code to reflect that.
Also, somehow the index.hbs files got overwritten with the plot point
stuff, don’t know how.